@@ -1,39 +1,29 @@
 package be.ictdynamic.ES_GEO_POC;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.Contact;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 /**
  * Created by wvdbrand on 31/03/2017.
  */
 @Configuration
-@EnableSwagger2
-@Import({springfox.documentation.spring.data.rest.configuration.SpringDataRestConfiguration.class})
+//@EnableSwagger2
+//@Import({springfox.documentation.spring.data.rest.configuration.SpringDataRestConfiguration.class})
 public class SwaggerConfig {
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("be.ictdynamic"))
-                .paths(PathSelectors.any())
-                .build()
-                .apiInfo(apiInfo());
-    }
-
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder()
-                .title("Elasticsearch GEO POC")
-                .description("This project has been designed/developed as a POC. Focus of the POC is to test a number of GEO related queries against Elasticsearch (ES).")
-                .contact(new Contact("Wim Van den Brande", "", "info@ictdynamic.be"))
-                .build();
-    }
+//    @Bean
+//    public Docket api() {
+//        return new Docket(DocumentationType.SWAGGER_2)
+//                .select()
+//                .apis(RequestHandlerSelectors.basePackage("be.ictdynamic"))
+//                .paths(PathSelectors.any())
+//                .build()
+//                .apiInfo(apiInfo());
+//    }
+//
+//    private ApiInfo apiInfo() {
+//        return new ApiInfoBuilder()
+//                .title("ElasticSearch GEO POC")
+//                .description("This project has been designed/developed as a POC. Focus of the POC is to test a number of GEO related queries against ElasticSearch (ES).")
+//                .contact(new Contact("Wim Van den Brande", "", "info@ictdynamic.be"))
+//                .build();
+//    }
 }
